Move update schedule to Entity class
